+++ Cards/community_cards.py
@@ -1,0 +1,13 @@
+# Definition of the CommunityCards class
+class CommunityCards:
+    def __init__(self):
+        self.cards = []
+
+    def add_card(self, card):
+        self.cards.append(card)
+
+    def reset(self):
+        self.cards = []
+
+    def __repr__(self):
+        return f"Community Cards: {self.cards}"
